+++ exchangelib/__init__.py
@@ -6,7 +6,7 @@ from .autodiscover import discover
 from .configuration import Configuration
 from .credentials import DELEGATE, IMPERSONATION, Credentials, ServiceAccount
 from .ewsdatetime import EWSDate, EWSDateTime, EWSTimeZone, UTC, UTC_NOW
-from .extended_properties import ExtendedProperty, ExternId
+from .extended_properties import ExtendedProperty
 from .folders import Folder, FolderCollection, SHALLOW, DEEP
 from .items import AcceptItem, TentativelyAcceptItem, DeclineItem, CalendarItem, CancelCalendarItem, Contact, \
     DistributionList, Message, PostItem, Task
@@ -43,12 +43,3 @@ def close_connections():
     from .protocol import close_connections as close_protocol_connections
     close_autodiscover_connections()
     close_protocol_connections()
-
-
-# Pre-register these extended properties. They are not part of the standard EWS fields but are useful for identification
-# when item originates in an external system.
-
-CalendarItem.register('extern_id', ExternId)
-Message.register('extern_id', ExternId)
-Contact.register('extern_id', ExternId)
-Task.register('extern_id', ExternId)